database负责转发

+++ MYClientDatabase/Assets/clientdatabase.sql
@@ -28,5 +28,5 @@ CREATE TABLE "tb_message" (
     "sendStatus" integer, -- 发送状态 0 发送中 1 已发送 2 发送失败
     "readList" TEXT, -- 使用逗号分割已读人员
     "kReadList" TEXT, -- 已读列表
-    "timestamp" INTEGER NOT null default 0
+    "timestamp" INTEGER(8) NOT null default 0
 );
